BatedorDePonto Controller Almost Ready, Some Changes in entity BatedorDePonto

+++ src/main/java/com/devPontes/oMordomo/services/BatedorDePontoServices.java
@@ -13,7 +13,7 @@ public interface BatedorDePontoServices {
     BatedorDePontoDTO registrarPontoFuncionario(PontoDTO ponto, Long funcionarioId, Long batedorId) throws Exception;
     BatedorDePontoDTO atualizarPontoFuncionario(Long pontoId, Long funcionarioId) throws Exception;
     BatedorDePontoDTO registrarFaltaFuncionario(LocalDate diaDaFalta, Long funcionarioId) throws Exception;
-    Long calcularHorasFuncionarioMes(Long funcionarioId) throws Exception;
+    Integer calcularHorasFuncionarioMes(Long funcionarioId) throws Exception;
     Double calcularBonusSalario(Long funcionarioId, Double reajuste) throws Exception;
     PontoDTO exibirPontoFuncionario(Long funcionarioId) throws Exception;
 }
